+++ examples/default/variables.tf
@@ -17,8 +17,3 @@ variable "stack" {
   description = "Project stack name"
   type        = string
 }
-
-variable "subnet_id" {
-  description = "EventHub subnet"
-  type        = string
-}