+++ examples/main/modules.tf
@@ -39,32 +39,23 @@ module "eventhub" {
 
   resource_group_name = module.rg.resource_group_name
 
-  eventhub_namespaces_hubs = {
-    # You can just create a eventhub_namespace
-    eventhub0 = {}
+  create_dedicated_cluster = true
 
-    # Or create a eventhub_namespace with some hubs with default values
-    eventhub1 = {
-      hubs = {
-        hub1 = {}
-        hub2 = {}
-      }
-    }
+  namespace_parameters = {
+    sku      = "Standard"
+    capacity = 2
+  }
 
-    eventhub2 = {
-      custom_name          = "testeventhub"
-      sku                  = "Standard"
-      capacity             = 1
-      auto_inflate_enabled = true
-      reader               = true
-      hubs = {
-        hubcentdeux = {
-          message_retention = 7
-          partition_count   = 2
-          sender            = true
-          manage            = true
-        }
-      }
+  network_rules_enabled = true
+  allowed_cidrs         = ["1.1.1.1/32"]
+  allowed_subnet_ids = [
+    var.subnet_id
+  ]
+
+  hubs_parameters = {
+    main = {
+      custom_name     = "main-queue-hub"
+      partition_count = 2
     }
   }