+++ variables-naming.tf
@@ -12,7 +12,14 @@ variable "name_suffix" {
 }
 
 variable "use_caf_naming" {
-  description = "Use the Azure CAF naming provider to generate default resource name. `custom_name` override this if set. Legacy default name is used if this is set to `false`."
+  description = "Use the Azure CAF naming provider to generate default resource name. `custom_namespace_name` override this if set. Legacy default name is used if this is set to `false`."
   type        = bool
   default     = true
+}
+
+# Custom resource names
+variable "custom_namespace_name" {
+  description = "Custom resource name for EventHub namespace."
+  type        = string
+  default     = ""
 }